remove deprecated tests
