+++ runtime/integration-tests/src/lib.rs
@@ -17,7 +17,6 @@
 mod evm;
 mod generic;
 mod liquidity_pools;
-mod pools;
 mod rewards;
 mod runtime_apis;
 mod utils;
