+++ runtime/integration-tests/src/utils/env.rs
@@ -935,3 +935,34 @@ pub fn pass_n(env: &mut TestEnv, n: u64) -> Result<(), ()> {
 
 	Ok(())
 }
+
+mod tests {
+	use super::*;
+
+	#[tokio::test]
+	async fn env_works() {
+		let mut env = test_env_default(Handle::current());
+
+		// FIXME: https://github.com/centrifuge/centrifuge-chain/issues/1219
+		// Breaks on >= 10 for fast-runtime since session length is 5 blocks
+		#[cfg(feature = "fast-runtime")]
+		let num_blocks = 9;
+		#[cfg(not(feature = "fast-runtime"))]
+		let num_blocks = 10;
+		let block_before = env
+			.with_state(Chain::Para(PARA_ID), || {
+				frame_system::Pallet::<Runtime>::block_number()
+			})
+			.expect("Cannot create block before");
+
+		frame_support::assert_ok!(pass_n(&mut env, num_blocks));
+
+		let block_after = env
+			.with_state(Chain::Para(PARA_ID), || {
+				frame_system::Pallet::<Runtime>::block_number()
+			})
+			.expect("Cannot create block after");
+
+		assert_eq!(block_before + num_blocks as u32, block_after)
+	}
+}
