+++ runtime/integration-tests/src/utils/mod.rs
@@ -20,9 +20,7 @@ pub mod evm;
 pub mod extrinsics;
 pub mod genesis;
 pub mod liquidity_pools_gateway;
-pub mod loans;
 pub mod logs;
-pub mod pools;
 pub mod preimage;
 pub mod time;
 pub mod tokens;
